rename local_state to zalsa_local
