+++ src/function/inputs.rs
@@ -1,4 +1,4 @@
-use crate::{local_state::QueryOrigin, Id};
+use crate::{zalsa_local::QueryOrigin, Id};
 
 use super::{Configuration, IngredientImpl};
