+++ src/tracked_struct/tracked_field.rs
@@ -91,7 +91,7 @@ where
         field_changed_at > revision
     }
 
-    fn origin(&self, _key_index: crate::Id) -> Option<crate::local_state::QueryOrigin> {
+    fn origin(&self, _key_index: crate::Id) -> Option<crate::zalsa_local::QueryOrigin> {
         None
     }
